+++ include/fem-dem/cfd_dem_coupling.h
@@ -194,10 +194,12 @@ private:
   dem_contact_build(unsigned int counter);
 
 
-  unsigned int                               coupling_frequency;
-  Tensor<1, 3>                               g;
-  std::vector<Tensor<1, 3>>                  torque;
-  std::vector<Tensor<1, 3>>                  force;
+  unsigned int coupling_frequency;
+  Tensor<1, 3> g;
+  ParticleInteractionOutcomes<DEM::CFDDEMProperties::PropertiesIndex>
+                                             contact_outcome;
+  std::vector<Tensor<1, 3>>                 &torque = contact_outcome.torque;
+  std::vector<Tensor<1, 3>>                 &force  = contact_outcome.force;
   std::vector<double>                        displacement;
   std::vector<double>                        MOI;
   double                                     neighborhood_threshold_squared;