+++ tests/dem/particle_particle_thermal_resistances.cc
@@ -115,18 +115,16 @@ test()
   pit_2->get_properties()[PropertiesIndex::specific_heat] = specific_heat;
 
   // Initializing variables
-  std::vector<Tensor<1, 3>> torque;
-  std::vector<Tensor<1, 3>> force;
-  std::vector<double>       MOI;
-  std::vector<double>       heat_transfer_rate;
+  ParticleInteractionOutcomes<PropertiesIndex> contact_outcome;
+  std::vector<double>                          MOI;
 
   particle_handler.sort_particles_into_subdomains_and_cells();
-  force.resize(particle_handler.get_max_local_particle_index());
-  torque.resize(force.size());
-  MOI.resize(force.size());
+  const unsigned int number_of_particles =
+    particle_handler.get_max_local_particle_index();
+  contact_outcome.resize_interaction_containers(number_of_particles);
+  MOI.resize(number_of_particles);
   for (auto &moi_val : MOI)
     moi_val = 1;
-  heat_transfer_rate.resize(force.size());
 
   contact_manager.update_local_particles_in_cells(particle_handler);
 
@@ -146,17 +144,14 @@ test()
       hertz_mindlin_limit_overlap,
     Parameters::Lagrangian::RollingResistanceMethod::constant_resistance>
     nonlinear_force_object(dem_parameters);
-  nonlinear_force_object
-    .calculate_particle_particle_contact_force_and_heat_transfer_rate(
-      contact_manager.get_local_adjacent_particles(),
-      contact_manager.get_ghost_adjacent_particles(),
-      contact_manager.get_local_local_periodic_adjacent_particles(),
-      contact_manager.get_local_ghost_periodic_adjacent_particles(),
-      contact_manager.get_ghost_local_periodic_adjacent_particles(),
-      dt,
-      torque,
-      force,
-      heat_transfer_rate);
+  nonlinear_force_object.calculate_particle_particle_contact(
+    contact_manager.get_local_adjacent_particles(),
+    contact_manager.get_ghost_adjacent_particles(),
+    contact_manager.get_local_local_periodic_adjacent_particles(),
+    contact_manager.get_local_ghost_periodic_adjacent_particles(),
+    contact_manager.get_ghost_local_periodic_adjacent_particles(),
+    dt,
+    contact_outcome);
 
   // Calculating additional parameters for thermal DEM
   auto     particle_one          = particle_handler.begin();
@@ -176,7 +171,8 @@ test()
   const double normal_overlap =
     (radius_one + radius_two) -
     particle_one_location.distance(particle_two_location);
-  const double normal_force_norm = force[particle_one->get_id()].norm();
+  const double normal_force_norm =
+    contact_outcome.force[particle_one->get_id()].norm();
   const double prandtl_gas =
     dynamic_viscosity_gas * specific_heat_gas / thermal_conductivity_gas;
   const double gas_parameter_m =